+++ src/main/java/Eclipse/Calculator/Input.java
@@ -5,18 +5,18 @@ import java.util.Scanner;
 public class Input {
 	
 	static Scanner in = new Scanner (System.in);
-			static double Number1;
-			static double Number2;
+			static int Number1;
+			static int Number2;
 			
-			public static Double input1 () {
+			public static int input1 () {
 				
 	System.out.println("Enter First Number Please ");
-	return Number1 = in.nextDouble();
+	return Number1 = in.nextInt();
 			}
 			
-			public static Double input2 () {
+			public static int input2 () {
 	
 	System.out.println("Enter Second Number Please");
-	return Number2 = in.nextDouble();
+	return Number2 = in.nextInt();
 			}
 }
