+++ src/test/java/Eclipse/Calculator/CalculatorTest.java
@@ -5,17 +5,19 @@ import org.junit.*;
 
 public class CalculatorTest {
 
+
+	
 	@Test
 	public void test_calculate_add_numbers() {
-		double number1 = 0;
-		double number2 = 1;
+		int number1 = 0;
+		int number2 = 1;
 
 		// expected result
 
-		double expected = 0;
+		int expected = 0;
 
 		// actual value
-		double actual = calculations.Division(number1, number2);
+		int actual = calculations.Division(number1, number2);
 		Assert.assertEquals(expected, actual, 0.1);
 	}
 }
